+++ apps/web/dao/Log.php
@@ -1,0 +1,10 @@
+<?php
+namespace WDAO;
+
+class Log extends \MDAOBASE\DaoBase
+{
+    public function __construct($data)
+    {
+        parent::__construct($data);
+    }
+}
